Add files via upload
v1.01 updates:

SF = Short-form
LF = Long-form

Concept behind this is that the short form sheet is laid out where all of the individual lots are on individual sub sheets within the master excel file. This requires a binding function. For the long form sheet, all compiled data is laid out in one sheet.